requirements.txt
appdirs==1.4.4
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-extensions==3.0.9
django-heroku==0.3.1
django-import-export==2.4.0
django-libsass==0.8
django-multiselectfield==0.1.12
django-node-assets==0.9.7
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
psycopg2-binary==2.8.6
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==20.0.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0

webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
